// src/types/apps/chatTypes.ts
export type ProfileUserType = {
  profile_picture: string | undefined
  id: number
  role: string
  about: string
  avatar: string
  fullName: string
  status: StatusType
  settings: {
    isNotificationsOn: boolean
    isTwoStepAuthVerificationEnabled: boolean
  }
}

export type MsgFeedbackType = {
  isSent: boolean
  isSeen: boolean
  isDelivered: boolean
}

export type ChatType = {
  message: string
  senderId: number
  time: Date | string
  feedback: MsgFeedbackType
}

export type ChatsObj = {
  id: number
  userId: number
  chat: ChatType[]
  unseenMsgs: number
  lastMessage?: ChatType
}

export type ContactType = {
  profile_picture?: any
  id: number
  role: string
  about: string
  avatar?: string
  fullName: string
  status: StatusType
  avatarColor?: ThemeColor
}

export type ChatsArrType = {
  name: string | undefined
  id: number
  role: string
  about: string
  chat: ChatsObj
  profile_picture?: string
  fullName: string
  status: StatusType
  avatarColor?: ThemeColor
}

export type Participant = {
  id: string
  userId: string
  user: UserDataType
}

export type Chat = {
  id: string
  fullName: string
  name: string
  profile_picture: string
  avatarColor?: ThemeColor
  status: StatusType
  participants: Participant[]
  createdAt: Date
  messages: [] | any
}

export type SelectedChatType = null | {
  fullName: string
  blockedUserId: string
  name: string
  status: "BLOCK" | "UNBLOCK"
  role: any
  chat: ChatsObj
  contact: ChatsArrType
  participants: Participant[]
  isBlocked: boolean
  id?: string
}

export type ChatStoreType = {
  conversation: any
  chats: ChatsArrType[] | null
  contacts: ContactType[] | null
  userProfile: ProfileUserType | null
  selectedChat: SelectedChatType
  status?: 'idle' | 'pending' | 'success' | 'error'
}

export type SendMsgParamsType = {
  chat?: ChatsObj
  message: string
  contact?: ChatsArrType
}
// ...
export type SendMsgComponentType = {
  store: ChatStoreType
  dispatch: Dispatch<any>
  sendMsg: (params: SendMsgParamsType) => void
}